requirements.txt
requests
twitchio
python-dotenv